packages/rag_aas/pyproject.toml
[project]
name = "rag_aas"
version = "0.1.0"
description = "RAG as a Service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.5",
    "langchain-postgres>=0.0.12",
    "langchain>=0.3.7",
    "pgvector==0.2.5",
    "pydantic-settings>=2.6.1",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
]


[dependency-groups]
dev = [
    "pytest>=8.3.3",
    "ruff>=0.7.4",
]
